Workshopping The Website 1.2.2
In this commit I added some finishing touches to the home page and also changed the colour of the header and few text. I also Created the "Work with us" page and the "Contact us page" and made the site more responsive.
what's left to do not is the learn with us page majorly and then a few other changes before I round up the site.